- Fixed a small bug in the tracking camera where it would pass a float   into the first parameter of Accelate() but that accepted an int. - Increased the tilt speed in celtest. - Implemented a new continous transition interpolation feature in the   delegation camera. This makes for MUCH smoother camera movement. - Added a new cel.property.cont_trans_speed property to control this new   transition speed. - Extended the iPcDelegateCamera interface with methods to get and set   this new transition speed. Default is 3.

+++ cel/trunk/plugins/propclass/cameras/tracking/tracking.h
@@ -125,7 +125,7 @@ private:
 
   struct Accelerator
   {
-    void Accelerate (int direction, float elapsedsecs);
+    void Accelerate (float direction, float elapsedsecs);
     float topspeed;
     float speed, accel;
   } pan, tilt;
